Fix issue 93: TSVN fails to handle svn:externals correctly upon commit.
When collecting all the statuses, also check the uuid of each file/folder. Non-equal uuids mean we have an svn:external from another repository and can't commit those in one step. So exclude those with a different uuid than the first we got (usually that from the parent dir), but show the user a warning dialog telling that there are modifications in one or more svn:external folders which need a separate commit.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@853 cc8e0003-5e82-43e2-9771-63201f7d8d9c