Patch from Simon Large (slarge@blazepoint.co.uk):
This implements the colour interpolation scheme described on the user list. I have tested it with quite a few of the Win2K colour schemes and it seems to work in most cases. It does not work that well with the reverse colour high-contrast schemes because there is not a big difference between green-on-black and green-on-very-dark-grey. But there again, the scintilla bit ignores the colour scheme altogether.

I have also attached the makefile in case anyone else wants to have a go at building TBlame the hard way.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@1701 cc8e0003-5e82-43e2-9771-63201f7d8d9c