Improved shell-update mechanism.  Previously, shell-update requests were being done once-per-target.  This caused a bad interaction with the shell-extension status cache, where first update would load the cache, and subsequent updates would get stale, cached data.
I have added a named event to signal from TortoiseProc to the shellext, to tell it to invalidate its cache.  I have added a collection-class to TortoiseProc which queues-up shell-update requests until either a timeout expires, or the app exits, at which point all the updates are done in one go.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@2278 cc8e0003-5e82-43e2-9771-63201f7d8d9c