Show only those properties in the combobox which actually can be used on the selected item.
I.e. if a file is selected in the Subversion properties tab, then only add those properties which can be used on files, and if a folder is selected only show those which can be used on folders.
If more than one item (file or folder) is selected, all properties are added to the combobox.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@2627 cc8e0003-5e82-43e2-9771-63201f7d8d9c