Patch from Hans-Emil Skogh:
* Better color-variation
By changing luminosity and saturation in addition to hue the colors are easier to distinguish if there are many of them. The small downside is that all of them are not as candy-lookin any more.

* Stacked tooltips
If more datapoints end up on top of eachother in the line graph, the tooltip now shows all the datapoints instead of the first one found.

* Lower font size limit
Font sizes are not allowed to be smaller than readable.

* Auto adjusting legend
If the window gets small enough, it won’t bother to paint the legend. It gets too crowded to be useful anyway. The information in the tooltips is enough then.

* Stacked graph
The line and bar graph can now be generated as stacked by passing “true” as the second parameter to SetGraphType(GraphType eType, bool bStackedGraph). (Bitte Lübbe! ;-) )

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@5271 cc8e0003-5e82-43e2-9771-63201f7d8d9c