Instead of asking the user for the client certificate file, show the Windows client cert dialog and get the certificate from the certificate store instead. Users will have to import their certificates into the cert store, but that's not a problem and is more secure anyway.
The index of the selected certificate is stored in the registry and reused. If the retry count is more than one, the user has to select the certificate again, since this is an indicator that the authentication with the stored index failed.
Only the index in the cert store is saved in the registry, so no confidential information is left on the system unencrypted.

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@21544 cc8e0003-5e82-43e2-9771-63201f7d8d9c