Add an advanced setting MergeAllowMixedRevisionsDefault to set the default value of the Allow mixed revisions (not recommended) checkbox in Merge options.
If not set, it still defaults to FALSE.

Never said this was a good idea but people thought otherwise. Quoting from the SVN book:

[[[
Without going into too much detail, this is because of limitations in the way merges are tracked by the svn:mergeinfo property (see the section called “Mergeinfo and Previews” for details). These limitations mean that merges into mixed-revision working copies can result in unexpected text and tree conflicts.[34] We don't want any needless conflicts, so we update the working copy and then reattempt the merge.
]]]


git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@29597 cc8e0003-5e82-43e2-9771-63201f7d8d9c